Simplify makefile using variables and implicit rules
The previous version of the makefile list all the object files multiple times.
To mitigate the potential oversight of updating every listed component
when making modifications to the makefile, and to streamline its structure,
we introduced variables (e.g., "OBJS") to define a text string, which is
subsequently substituted in various instances throughout the file.

Since make can automatically deduce the implicit rule for updating
a '.o' file from a corresponding '.c' file using a 'cc -c' command,
it is advisable to exclude the recipes from the rules for the object
files.